fix(runner): platform-split ModeDaemon startup for Windows
Two Windows-specific regressions surfaced when CI actually ran the new
Windows e2e tests:

1. TestStartDaemonWindows (poddaemon unit suite) hit
   "processmgr: Init has not been called" because the unit suite has no
   TestMain on Windows. Added testmain_windows_test.go with the same
   processmgr.Init + LauncherSubcommand wiring used by the integration
   suite.

2. TestModeDaemon_Windows_StartStop failed with
   "fork/exec ...: not supported by windows". Root cause: Go's os/exec
   does not inherit ExtraFiles on Windows, so the launcher's fd-3 PID
   pipe scheme is structurally broken there. Fix: split startDaemon by
   platform — Unix keeps the double-fork launcher (still required for
   zombie-free detachment), Windows spawns the daemon directly with
   DETACHED_PROCESS + CREATE_NEW_PROCESS_GROUP, which already provides
   the same parent-detachment guarantees Windows users need (Windows has
   no zombie state).

process_daemon.go now holds only the cross-platform daemonProcess type
and monitorLoop; startDaemon + readDaemonPID live in the *_unix.go file,
and process_daemon_windows.go grows its own startDaemon that skips the
launcher path entirely.
